app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cookiecutter==2.6.0
cycler==0.12.1
fastapi==0.115.14
Flask==3.1.1
Flask-JWT-Extended==4.7.1
flaskwebgui==1.1.6
fonttools==4.58.5
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psutil==6.1.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
waitress==3.0.2
watchdog==4.0.2
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
